Apply VZ exclusion to a data-frame containing list-columns of yi subsets

Apply VZ exclusion to a data-frame containing list-columns of yi subsets

Usage

apply_VZ_exclusions(df = data.frame(), VZ_colname, VZ_cutoff)

Arguments

df

A dataframe of yi data subsets generated by generate_yi_subsets() or split_yi_subsets().

VZ_colname

Either A character vector of length 1, the name of the column in the dataframes stored in df's list-column data that contains the VZ values. Or else a named list of character values, where the names are the dataset names and the values are the VZ_colnames for each dataset in df.

VZ_cutoff

A numeric vector of length 1, values equal to or greater than this value of VZ will be filtered out of the dataframes stored in df's list-column data, else a named list of numeric values, where the names are the dataset names and the values are the VZ_cutoffs for each dataset in df.

Value

A dataframe of yi subsets, whose extreme values of VZ have been removed.

Details

df must contain the columns "data", "diversity_data" and "dataset". If only one value of VZ_colname and VZ_cutoff is supplied, it will be recycled to match the number of datasets in df.

If a named list is supplied for VZ_colname and VZ_cutoff, the names must match the dataset names in df.

See also

Examples

data(ManyEcoEvo_yi)
ManyEcoEvo_yi %>%
  prepare_response_variables(
    estimate_type = "yi",
    param_table =
      ManyEcoEvo:::analysis_data_param_tables,
    dataset_standardise = "blue tit",
    dataset_log_transform = "eucalyptus") %>%
  generate_yi_subsets() %>% 
  apply_VZ_exclusions(VZ_colname = 
                        list("eucalyptus" = "se_log", 
                             "blue tit" = "VZ"), 
                      VZ_cutoff = 3)
#> 
#> ── Generating out-of-sample prediction subsets. ────────────────────────────────
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ✔ Applied back-transformation for squared effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for squared effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for squared effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for squared effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for squared effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for squared effect sizes or out-of-sample predictions.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ✔ Applied back-transformation for ^100 effect sizes or out of sample predictions.
#> ✔ Applied back-transformation for cubed effect sizes
#> ✔ Applied back-transformation for ^100 effect sizes or out of sample predictions.
#> ✔ Applied back-transformation for cubed effect sizes
#> ✔ Applied back-transformation for ^100 effect sizes or out of sample predictions.
#> ✔ Applied back-transformation for cubed effect sizes
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ✔ Applied back-transformation for ^14 effect sizes or out of sample predictions.
#> ✔ Applied back-transformation for ^14 effect sizes or out of sample predictions.
#> ✔ Applied back-transformation for ^14 effect sizes or out of sample predictions.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ℹ No back-transformation required, identity link used.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for square-root transformed effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for square-root transformed effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for square-root transformed effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ℹ Standardising and/or log-transforming response variables for "yi" estimates.
#> 
#> ── Computing meta-analysis inputsfor `estimate_type` = "yi" ────────────────────
#> 
#> ── Standardising out-of-sample predictions ──
#> 
#> ── Computing meta-analysis inputs: ─────────────────────────────────────────────
#> 
#> ── Log-transforming response-variable ──
#> 
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> 
#> ── Applying VZ exclusions ──────────────────────────────────────────────────────
#> ! `VZ_cutoff` = 3 was recycled to match the number of unique datasets in `df`.
#> 
#> ── Excluding extreme values of VZ ──
#> 
#> → 0 extreme values of `VZ` removed at threshold of 3 for `dataset` "blue tit", `estimate_type` = "y25".
#> 
#> ── Excluding extreme values of VZ ──
#> 
#> → 0 extreme values of `VZ` removed at threshold of 3 for `dataset` "blue tit", `estimate_type` = "y25".
#> 
#> ── Excluding extreme values of VZ ──
#> 
#> → 0 extreme values of `VZ` removed at threshold of 3 for `dataset` "blue tit", `estimate_type` = "y25".
#> 
#> ── Excluding extreme values of VZ ──
#> 
#> → 0 extreme values of `se_log` removed at threshold of 3 for `dataset` "blue tit", `estimate_type` = "y25".
#> 
#> ── Excluding extreme values of VZ ──
#> 
#> → 0 extreme values of `se_log` removed at threshold of 3 for `dataset` "blue tit", `estimate_type` = "y25".
#> 
#> ── Excluding extreme values of VZ ──
#> 
#> → 0 extreme values of `se_log` removed at threshold of 3 for `dataset` "blue tit", `estimate_type` = "y25".
#> # A tibble: 6 × 4
#>   dataset    estimate_type data               diversity_data    
#>   <chr>      <chr>         <list>             <named list>      
#> 1 blue tit   y25           <tibble [65 × 26]> <tibble [65 × 54]>
#> 2 blue tit   y50           <tibble [60 × 26]> <tibble [60 × 54]>
#> 3 blue tit   y75           <tibble [65 × 26]> <tibble [65 × 54]>
#> 4 eucalyptus y25           <tibble [49 × 27]> <tibble [49 × 61]>
#> 5 eucalyptus y50           <tibble [51 × 27]> <tibble [51 × 61]>
#> 6 eucalyptus y75           <tibble [52 × 27]> <tibble [52 × 61]>