Back Transform Response Variables - yi

Back Transform Response Variables - yi

Usage

back_transform_response_vars_yi(
  dat,
  estimate_type = character(1L),
  dataset = character(1L)
)

Arguments

dat

A dataframe of out of sample predictions analyst submission data

estimate_type

The type of estimate to be standardised. Character vector of length 1, whose value may be "yi", "y25", "y50", "y75".

dataset

One of either "blue tit" or "eucalyptus"

Value

A tibble of analyst data with standardised values contained in a list-column called 'back_transformed_data'

See also