Compute meta-analysis inputs for a nested-dataframe containing different datasets/subsets of analyst data

Computes the sorensen diversity indices and joins it to the prepared data in preparation for meta-analysing all subsets of data with meta_analyse_datasets().

Usage

compute_MA_inputs(ManyEcoEvo, estimate_type = NULL)

Arguments

ManyEcoEvo

A dataframe grouped by the character columns dataset, estimate_type, exclusion_set. Each group corresponds to a subset of the full dataset, and has the subset analyst data stored in data, with its corresponding subset diversity_data.

estimate_type

character string, one of "Zr", "yi", "y25", "y50", "y75".

Value

A dataframe that includes the additional columns in ManyEcoEvo, but with added columns diversity_indices and effects_analysis.

Details

The name of the subset is derived from the funs generated by subset_fns_yi and/or subset_fns_Zr.

Computes sorensen diversity indices diversity_indices for each subset of data returning them in the list-columns diversity_indices and joins them to the subset analyst data, returning it in the list-column effects_analysis.

See also