Fit model of Box-Cox transformed deviation scores as a function random-effects inclusion in analyses

Fits a univariate glm of Box-Cox transformed absolute deviation from the meta-analytic mean scores as a function of whether the analysis was a mixed effects model (i.e. included random effects) or not.

Usage

fit_uni_mixed_effects(data, N = 5)

Arguments

data

Dataframe containing Box-Cox transformed absolute deviation scores and binary column called mixed_model describing whether or not the analysis used a mixed-effects model.

N

threshold number of analyses in each predictor category for fitting model

Value

A fitted model object of class glm and parsnip

See also

Examples

# library(tidyverse);library(targets);library(metafor);library(tidymodels)
# tar_load(meta_analysis_outputs)
# fit_uni_mixed_effects(meta_analysis_results$data[[1]])