Make visualisations wrapper function

Compute model summaries, tidy model summaries, model fit statistics, funnel plots and forest plots for a tibble of multiple fitted models

Usage

make_viz(data)

Arguments

data

a nested tibble with processed and standardised data stored in list-column data, grouped by variables exclusion_set, dataset, estimate_type, publishable_subset, expertise_subset, collinearity_subset. Each group contains a list-column model containing fitted models of class rma.uni, rma.mv or merMod.

Value

a nested tibble containing model summaries, fit statistics, plots, model parameters and other components of fitted model objects, see details.

Details

make_viz() is a wrapper function that takes a nested tibble of fitted models and computes model summaries, tidy model summaries, model fit statistics, funnel plots and forest plots for each model. The function is designed to be used in conjunction with ManyEcoEvo_results-type datasets, which contains multiple fitted models for each dataset and estimate type.

The following functions are applied to each model:

mod_summary: base::summary() to extract model summaries

tidy_mod_summary: broom.mixed::tidy() to extract tidy model summaries

mod_fit_stats / mod_glance: performance::performance() and broom::glance() to extract model fit statistics

funnel_plot: metaviz::viz_funnel() to create funnel plots for rma.uni models

forest_plot: gg_forest() to create forest plots for rma.mv models

MA_fit_stats: get_MA_fit_stats() to extract model fit statistics for rma.mv models

model_params: parameters::parameters() to extract model parameters

Note that where the fitted model object doesn't exist, i.e. is a NA or NULL value, the function will return NA for all components.

See also

Examples

make_viz(ManyEcoEvo_results)
#> Random effect variances not available. Returned R2 does not account for random effects.
#> Random effect variances not available. Returned R2 does not account for random effects.
#> Random effect variances not available. Returned R2 does not account for random effects.
#> Random effect variances not available. Returned R2 does not account for random effects.
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "blue tit" ──
#> 
#> ── Creating gg-forest-plot of `estimate_type` estimates for `dataset` = "eucalyptus" ──
#> 
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> Package 'merDeriv' needs to be installed to compute confidence intervals
#>   for random effect parameters.
#> # A tibble: 133 × 16
#>    exclusion_set dataset publishable_subset expertise_subset collinearity_subset
#>    <chr>         <chr>   <chr>              <chr>            <chr>              
#>  1 complete      blue t… All                All              All                
#>  2 complete      blue t… All                All              All                
#>  3 complete      blue t… All                All              All                
#>  4 complete      blue t… All                All              All                
#>  5 complete      blue t… All                All              All                
#>  6 complete      blue t… All                All              All                
#>  7 complete      blue t… All                All              All                
#>  8 complete      eucaly… All                All              All                
#>  9 complete      eucaly… All                All              All                
#> 10 complete      eucaly… All                All              All                
#> # ℹ 123 more rows
#> # ℹ 11 more variables: estimate_type <chr>, model_name <chr>, model <list>,
#> #   mod_summary <list>, tidy_mod_summary <list>, mod_fit_stats <list>,
#> #   mod_glance <list>, funnel_plot <list>, forest_plot <list>,
#> #   MA_fit_stats <list>, model_params <list>