Split data frame by groups and name elements

Split data frame by groups and name elements

Usage

named_group_split(.data, grouping_variable)

Arguments

.data

A tbl.

grouping_variable

Unquoted variable name to group columns by

Value

A named list of tibbles. Each tibble contains the rows of .tbl for the associated group and all the columns, including the grouping variables. Note that this returns a list_of which is slightly stricter than a simple list but is useful for representing lists where every element has the same type.

Details

Function will fail if character string is provided to grouping_variable instead of bare variable name.

Examples

named_group_split(ManyEcoEvo::euc_data , Property)
#> $Barlow
#> # A tibble: 18 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 1        2006-07-24 Winter 2006            1  310042  5934571 nne   
#>  2 2        2006-07-24 Winter 2006            2  310087  5934567 ne    
#>  3 3        2006-07-24 Winter 2006            3  310240  5934699 se    
#>  4 4        2006-07-24 Winter 2006            4  309998  5934968 0     
#>  5 5        2006-07-24 Winter 2006            5  309958  5934752 0     
#>  6 6        2006-07-24 Winter 2006            6  310137  5934765 0     
#>  7 118      2006-11-10 Spring 2006            1  310059  5935151 e     
#>  8 119      2006-11-10 Spring 2006            2  310272  5935064 0     
#>  9 120      2006-11-10 Spring 2006            3  310195  5935288 n     
#> 10 121      2006-11-10 Spring 2006            4  310218  5934982 ne    
#> 11 122      2006-11-10 Spring 2006            5  310034  5934700 e     
#> 12 123      2006-11-10 Spring 2006            6  309996  5934758 n     
#> 13 235      2007-04-19 Autumn 2007            1  310163  5935181 e     
#> 14 236      2007-04-19 Autumn 2007            2  310038  5935014 s     
#> 15 237      2007-04-19 Autumn 2007            3  310122  5934637 0     
#> 16 238      2007-04-19 Autumn 2007            4  310161  5934480 0     
#> 17 239      2007-04-19 Autumn 2007            5  310266  5934597 se    
#> 18 240      2007-04-19 Autumn 2007            6  310280  5935010 0     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $Blaber
#> # A tibble: 15 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 7        2006-07-25 Winter 2006            1  386643  5936462 e     
#>  2 8        2006-07-25 Winter 2006            2  386736  5936384 e     
#>  3 9        2006-07-25 Winter 2006            3  386732  5936315 e     
#>  4 10       2006-07-25 Winter 2006            4  386633  5936324 e     
#>  5 11       2006-07-25 Winter 2006            5  386896  5936282 e     
#>  6 124      2006-11-14 Spring 2006            1  387003  5936393 e     
#>  7 125      2006-11-14 Spring 2006            2  386476  5936219 e     
#>  8 126      2006-11-14 Spring 2006            3  386827  5936476 e     
#>  9 127      2006-11-14 Spring 2006            4  386616  5936267 e     
#> 10 128      2006-11-14 Spring 2006            5  386709  5936234 e     
#> 11 241      2007-04-17 Autumn 2007            1  386272  5936457 e     
#> 12 242      2007-04-17 Autumn 2007            2  387031  5936429 e     
#> 13 243      2007-04-17 Autumn 2007            3  386358  5936349 e     
#> 14 244      2007-04-17 Autumn 2007            4  386353  5936444 e     
#> 15 245      2007-04-17 Autumn 2007            5  386675  5936398 e     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $DClark
#> # A tibble: 18 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 12       2006-07-19 Winter 2006            1  380890  5959540 se    
#>  2 13       2006-07-19 Winter 2006            2  380734  5959474 s     
#>  3 14       2006-07-19 Winter 2006            3  381469  5959397 e     
#>  4 15       2006-07-19 Winter 2006            4  381024  5959384 s     
#>  5 16       2006-07-19 Winter 2006            5  380890  5959591 se    
#>  6 17       2006-07-19 Winter 2006            6  381203  5959676 e     
#>  7 129      2006-11-14 Spring 2006            1  380818  5959432 nw    
#>  8 130      2006-11-14 Spring 2006            2  381282  5959592 s     
#>  9 131      2006-11-14 Spring 2006            3  381475  5959610 e     
#> 10 132      2006-11-14 Spring 2006            4  381423  5959607 e     
#> 11 133      2006-11-14 Spring 2006            5  381170  5959382 s     
#> 12 134      2006-11-14 Spring 2006            6  381127  5959717 e     
#> 13 246      2007-04-17 Autumn 2007            1  381391  5959468 s     
#> 14 247      2007-04-17 Autumn 2007            2  380947  5959427 0     
#> 15 248      2007-04-17 Autumn 2007            3  381324  5959557 0     
#> 16 249      2007-04-17 Autumn 2007            4  380756  5959400 e     
#> 17 250      2007-04-17 Autumn 2007            5  381203  5959726 s     
#> 18 251      2007-04-17 Autumn 2007            6  381236  5959842 e     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $Gough
#> # A tibble: 27 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 18       2006-07-26 Winter 2006            1  358873  5921525 0     
#>  2 19       2006-07-26 Winter 2006            2  359203  5921862 0     
#>  3 20       2006-07-26 Winter 2006            3  358793  5921874 se    
#>  4 21       2006-07-26 Winter 2006            4  358452  5921673 n     
#>  5 22       2006-07-26 Winter 2006            5  359497  5922192 s     
#>  6 23       2006-07-26 Winter 2006            6  359480  5955191 n     
#>  7 24       2006-07-26 Winter 2006            7  359428  5921795 s     
#>  8 25       2006-07-26 Winter 2006            8  358742  5921709 s     
#>  9 26       2006-07-26 Winter 2006            9  359198  5921949 s     
#> 10 135      2006-11-04 Spring 2006            1  358522  5921752 s     
#> # ℹ 17 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $Green
#> # A tibble: 12 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 27       2006-07-27 Winter 2006            1  360032  5871750 n     
#>  2 28       2006-07-27 Winter 2006            2  359983  5871719 n     
#>  3 29       2006-07-27 Winter 2006            3  360061  5871604 e     
#>  4 30       2006-07-27 Winter 2006            4  359779  5871566 e     
#>  5 144      2006-12-07 Spring 2006            1  359810  5871505 w     
#>  6 145      2006-12-07 Spring 2006            2  359692  5871634 n     
#>  7 146      2006-12-07 Spring 2006            3  359865  5871733 s     
#>  8 147      2006-12-07 Spring 2006            4  359972  5871655 w     
#>  9 261      2007-05-09 Autumn 2007            1  359972  5871635 n     
#> 10 262      2007-05-09 Autumn 2007            2  359924  5871773 n     
#> 11 263      2007-05-09 Autumn 2007            3  359869  5871765 n     
#> 12 264      2007-05-09 Autumn 2007            4  359715  5871694 e     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $Hawkey
#> # A tibble: 12 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 31       2006-07-25 Winter 2006            1  330170  6010619 0     
#>  2 32       2006-07-25 Winter 2006            2  330187  6010583 0     
#>  3 33       2006-07-25 Winter 2006            3  330110  6010544 0     
#>  4 34       2006-07-25 Winter 2006            4  330200  6010497 0     
#>  5 148      2006-11-10 Spring 2006            1  330118  6010800 0     
#>  6 149      2006-11-10 Spring 2006            2  330540  6010844 0     
#>  7 150      2006-11-10 Spring 2006            3  330210  6010800 0     
#>  8 151      2006-11-10 Spring 2006            4  330125  6010555 0     
#>  9 265      2007-04-18 Autumn 2007            1  330074  6010835 0     
#> 10 266      2007-04-18 Autumn 2007            2  330228  6010621 0     
#> 11 267      2007-04-18 Autumn 2007            3  330258  6010626 0     
#> 12 268      2007-04-18 Autumn 2007            4  330133  6010679 0     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $JClark
#> # A tibble: 21 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 35       2006-07-26 Winter 2006            1  367739  5821489 sw    
#>  2 36       2006-07-26 Winter 2006            2  368058  5872589 w     
#>  3 37       2006-07-26 Winter 2006            3  368143  5872626 s     
#>  4 38       2006-07-26 Winter 2006            4  368131  5872189 s     
#>  5 39       2006-07-26 Winter 2006            5  367696  5872366 nw    
#>  6 40       2006-07-26 Winter 2006            6  368650  5872203 w     
#>  7 41       2006-07-26 Winter 2006            7  368188  5872013 se    
#>  8 152      2006-12-07 Spring 2006            1  367727  5872320 s     
#>  9 153      2006-12-07 Spring 2006            2  367851  5872147 nw    
#> 10 154      2006-12-07 Spring 2006            3  367744  5871824 w     
#> # ℹ 11 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $Kellock
#> # A tibble: 27 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 42       2006-07-26 Winter 2006            1  409176  5987415 n     
#>  2 43       2006-07-25 Winter 2006            2  409242  5987228 s     
#>  3 44       2006-07-25 Winter 2006            3  408941  5987483 e     
#>  4 45       2006-07-26 Winter 2006            4  409111  5987866 s     
#>  5 46       2006-07-25 Winter 2006            5  409489  5987803 0     
#>  6 47       2006-07-25 Winter 2006            6  408888  5987382 se    
#>  7 48       2006-07-25 Winter 2006            7  409174  5987798 s     
#>  8 49       2006-07-25 Winter 2006            8  408864  5987984 w     
#>  9 50       2006-07-25 Winter 2006            9  409135  5987761 ne    
#> 10 159      2006-11-10 Spring 2006            1  408962  5987848 w     
#> # ℹ 17 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $Martin
#> # A tibble: 33 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 51       2006-07-20 Winter 2006            1  387782  5968745 e     
#>  2 52       2006-07-20 Winter 2006            2  388746  5967248 0     
#>  3 53       2006-07-20 Winter 2006            3  388765  5956890 ne    
#>  4 54       2006-07-20 Winter 2006            4  388358  5956799 0     
#>  5 55       2006-07-20 Winter 2006            5  388477  5956640 0     
#>  6 56       2006-07-20 Winter 2006            6  388029  5956943 0     
#>  7 57       2006-07-20 Winter 2006            7  388341  5956084 0     
#>  8 58       2006-07-20 Winter 2006            8  388423  5956006 se    
#>  9 59       2006-07-20 Winter 2006            9  388452  5971319 0     
#> 10 60       2006-07-20 Winter 2006           10  388732  5960630 0     
#> # ℹ 23 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $McCracken
#> # A tibble: 24 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 62       2006-07-18 Winter 2006            1  358110  5957069 0     
#>  2 63       2006-07-18 Winter 2006            2  358317  5957468 0     
#>  3 64       2006-07-18 Winter 2006            3  357720  5957153 0     
#>  4 65       2006-07-18 Winter 2006            4  358331  5957240 0     
#>  5 66       2006-07-18 Winter 2006            5  358632  5957291 0     
#>  6 67       2006-07-18 Winter 2006            6  358145  5957383 0     
#>  7 68       2006-07-18 Winter 2006            7  358137  5957448 0     
#>  8 69       2006-07-18 Winter 2006            8  358289  5956904 0     
#>  9 179      2006-11-11 Spring 2006            1  358300  5956890 0     
#> 10 180      2006-11-11 Spring 2006            2  358234  5957048 0     
#> # ℹ 14 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $Olive
#> # A tibble: 15 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 70       2006-07-27 Winter 2006            1  377590  5876368 0     
#>  2 71       2006-07-27 Winter 2006            2  377744  5976421 e     
#>  3 72       2006-07-27 Winter 2006            3  377219  5876496 s     
#>  4 73       2006-07-27 Winter 2006            4  377015  5876440 e     
#>  5 74       2006-07-27 Winter 2006            5  377790  5876366 e     
#>  6 187      2006-12-06 Spring 2006            1  377201  5876434 s     
#>  7 188      2006-12-06 Spring 2006            2  377301  5876368 s     
#>  8 189      2006-12-06 Spring 2006            3  377707  5876436 s     
#>  9 190      2006-12-06 Spring 2006            4  377587  5876420 n     
#> 10 191      2006-12-06 Spring 2006            5  377862  5876384 0     
#> 11 304      2007-05-08 Autumn 2007            1  377394  5876407 n     
#> 12 305      2007-05-08 Autumn 2007            2  377181  5876421 n     
#> 13 306      2007-05-08 Autumn 2007            3  377259  5876280 e     
#> 14 307      2007-05-08 Autumn 2007            4  377951  5876374 e     
#> 15 308      2007-05-08 Autumn 2007            5  377623  5876345 0     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $Rokahr
#> # A tibble: 21 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 75       2006-07-25 Winter 2006            1  387771  5985347 0     
#>  2 76       2006-07-25 Winter 2006            2  388366  5985356 0     
#>  3 77       2006-07-25 Winter 2006            3  387809  5985141 0     
#>  4 78       2006-07-25 Winter 2006            4  387671  5984967 0     
#>  5 79       2006-07-25 Winter 2006            5  387695  5985380 0     
#>  6 80       2006-07-25 Winter 2006            6  387753  5985068 0     
#>  7 81       2006-07-25 Winter 2006            7  388043  5985088 0     
#>  8 192      2006-11-10 Spring 2006            1  387721  5985370 0     
#>  9 193      2006-11-10 Spring 2006            2  388091  5985252 0     
#> 10 194      2006-11-10 Spring 2006            3  387748  5985167 0     
#> # ℹ 11 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $Sharrock
#> # A tibble: 12 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 82       2006-07-19 Winter 2006            1  383318  5955469 nw    
#>  2 83       2006-07-19 Winter 2006            2  383400  5955169 w     
#>  3 84       2006-07-19 Winter 2006            3  383228  5955360 s     
#>  4 85       2006-07-19 Winter 2006            4  383368  5955220 w     
#>  5 199      2006-11-14 Spring 2006            1  383259  5955400 w     
#>  6 200      2006-11-14 Spring 2006            2  383454  5955569 w     
#>  7 201      2006-11-14 Spring 2006            3  383340  5955291 nw    
#>  8 202      2006-11-14 Spring 2006            4  383247  5955254 nw    
#>  9 316      2007-04-17 Autumn 2007            1  383431  5955513 w     
#> 10 317      2007-04-17 Autumn 2007            2  383359  5955547 0     
#> 11 318      2007-04-17 Autumn 2007            3  383463  5955256 0     
#> 12 319      2007-04-17 Autumn 2007            4  383492  5955228 s     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $Staff
#> # A tibble: 27 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 86       2006-07-27 Winter 2006            1  385708  5939744 nw    
#>  2 87       2006-07-27 Winter 2006            2  386193  5939749 s     
#>  3 88       2006-07-27 Winter 2006            3  385599  5939313 0     
#>  4 89       2006-07-27 Winter 2006            4  385701  5939171 sw    
#>  5 90       2006-07-27 Winter 2006            5  386179  5939590 n     
#>  6 91       2006-07-27 Winter 2006            6  386037  5939836 n     
#>  7 92       2006-07-27 Winter 2006            7  385604  5939140 e     
#>  8 93       2006-07-27 Winter 2006            8  385700  5939655 s     
#>  9 94       2006-07-27 Winter 2006            9  385991  5939583 n     
#> 10 203      2006-11-15 Spring 2006            1  386235  5939864 nw    
#> # ℹ 17 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $Stoney
#> # A tibble: 21 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 95       2006-07-27 Winter 2006            1  410529  5895093 e     
#>  2 96       2006-07-27 Winter 2006            2  410683  5895072 e     
#>  3 97       2006-07-27 Winter 2006            3  410440  5895054 ne    
#>  4 98       2006-07-27 Winter 2006            4  410545  5895038 e     
#>  5 99       2006-07-27 Winter 2006            5  410374  5895094 se    
#>  6 100      2006-07-27 Winter 2006            6  410210  5895023 e     
#>  7 101      2006-07-27 Winter 2006            7  410284  5895820 ne    
#>  8 212      2006-12-08 Spring 2006            1  410616  5899138 e     
#>  9 213      2006-12-08 Spring 2006            2  410623  5899408 e     
#> 10 214      2006-12-08 Spring 2006            3  410572  5898758 e     
#> # ℹ 11 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
#> $Taylor
#> # A tibble: 15 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 102      2006-07-24 Winter 2006            1  389720  5954557 0     
#>  2 103      2006-07-24 Winter 2006            2  378805  5954456 0     
#>  3 104      2006-07-24 Winter 2006            3  389669  5954833 0     
#>  4 105      2006-07-24 Winter 2006            4  389746  5954624 0     
#>  5 106      2006-07-24 Winter 2006            5  389635  5954910 0     
#>  6 219      2006-10-14 Spring 2006            1  389718  5954830 0     
#>  7 220      2006-10-14 Spring 2006            2  389698  5954770 0     
#>  8 221      2006-10-14 Spring 2006            3  389775  5954594 0     
#>  9 222      2006-10-14 Spring 2006            4  389817  5954875 0     
#> 10 223      2006-10-14 Spring 2006            5  389750  5954882 0     
#> 11 336      2007-04-16 Autumn 2007            4  389851  5954634 0     
#> 12 337      2007-04-16 Autumn 2007            1  389643  5954703 0     
#> 13 338      2007-04-16 Autumn 2007            2  389653  5954793 0     
#> 14 339      2007-04-16 Autumn 2007            3  389607  5954928 0     
#> 15 340      2007-04-16 Autumn 2007            5  389854  5954441 0     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $Wakefield
#> # A tibble: 12 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 107      2006-07-27 Winter 2006            1  348307  5874042 nw    
#>  2 108      2006-07-27 Winter 2006            2  348206  5874084 s     
#>  3 109      2006-07-27 Winter 2006            3  348532  5874001 nw    
#>  4 110      2006-07-27 Winter 2006            4  348165  5874028 s     
#>  5 224      2006-12-07 Spring 2006            1  348305  5873844 s     
#>  6 225      2006-12-07 Spring 2006            2  348562  5874018 n     
#>  7 226      2006-12-07 Spring 2006            3  348253  5879750 s     
#>  8 227      2006-12-07 Spring 2006            4  348391  5873773 s     
#>  9 341      2007-05-09 Autumn 2007            1  348451  5873910 se    
#> 10 342      2007-05-09 Autumn 2007            2  348442  5873883 s     
#> 11 343      2007-05-09 Autumn 2007            3  348251  5873900 0     
#> 12 344      2007-05-09 Autumn 2007            4  348188  5873921 s     
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>,
#> #   BareGround_cover <dbl>, Litter_cover <dbl>, MossLichen_cover <dbl>, …
#> 
#> $Yorston
#> # A tibble: 21 × 42
#>    SurveyID Date       Season      `Quadrat no` Easting Northing Aspect
#>    <fct>    <date>     <chr>              <dbl>   <dbl>    <dbl> <chr> 
#>  1 111      2006-07-26 Winter 2006            1  359010  5894495 nw    
#>  2 112      2006-07-26 Winter 2006            2  359018  5894606 nw    
#>  3 113      2006-07-26 Winter 2006            3  358205  5895793 s     
#>  4 114      2006-07-26 Winter 2006            4  358932  5894600 nw    
#>  5 115      2006-07-26 Winter 2006            5  358861  5894634 w     
#>  6 116      2006-07-26 Winter 2006            6  358074  5895284 s     
#>  7 117      2006-07-26 Winter 2006            7  357961  5895704 ne    
#>  8 228      2006-11-17 Spring 2006            1  358010  5895886 w     
#>  9 229      2006-11-17 Spring 2006            2  358179  5895876 w     
#> 10 230      2006-11-17 Spring 2006            3  358917  5894850 w     
#> # ℹ 11 more rows
#> # ℹ 35 more variables: `Landscape position` <chr>,
#> #   ExoticAnnualGrass_cover <dbl>, ExoticAnnualHerb_cover <dbl>,
#> #   ExoticPerennialHerb_cover <dbl>, ExoticPerennialGrass_cover <dbl>,
#> #   ExoticShrub_cover <dbl>, NativePerennialFern_cover <dbl>,
#> #   NativePerennialGrass_cover <dbl>, NativePerennialHerb_cover <dbl>,
#> #   NativePerennialGraminoid_cover <dbl>, NativeShrub_cover <dbl>, …
#> 
named_group_split(ManyEcoEvo::blue_tit_data, hatch_Area)
#> $B
#> # A tibble: 726 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804103                 2001              201123 B100      .             
#>  2 P804104                 2001              201123 B100      .             
#>  3 P804106                 2001              201123 B100      .             
#>  4 P804107                 2001              201123 B100      .             
#>  5 P804108                 2001              201123 B100      .             
#>  6 P804110                 2001              201123 B100      .             
#>  7 P804112                 2001              201123 B100      .             
#>  8 P804115                 2001              201123 B100      .             
#>  9 P804117                 2001              201123 B100      .             
#> 10 P804118                 2001              201123 B100      .             
#> # ℹ 716 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $C
#> # A tibble: 577 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804016                 2001              201194 C130      P803008       
#>  2 P804017                 2001              201194 C130      P803008       
#>  3 P804018                 2001              201167 C26       P803005       
#>  4 P804019                 2001              201194 C130      P803008       
#>  5 P804020                 2001              201167 C26       P803005       
#>  6 P804021                 2001              201194 C130      P803008       
#>  7 P804022                 2001              201194 C130      P803008       
#>  8 P804023                 2001              201194 C130      P803008       
#>  9 P804024                 2001              201167 C26       P803005       
#> 10 P804025                 2001              201194 C130      P803008       
#> # ℹ 567 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $CP
#> # A tibble: 229 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804196                 2001              201225 CP40A     P803102       
#>  2 P804197                 2001              201225 CP40A     P803102       
#>  3 P804198                 2001              201225 CP40A     P803102       
#>  4 P804199                 2001              201225 CP40A     P803102       
#>  5 P804456                 2001              201225 CP40A     P803102       
#>  6 P804457                 2001              201225 CP40A     P803102       
#>  7 P804458                 2001              201225 CP40A     P803102       
#>  8 P804459                 2001              201225 CP40A     P803102       
#>  9 P804463                 2001              201225 CP40A     P803102       
#> 10 P804464                 2001              201225 CP40A     P803102       
#> # ℹ 219 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $E
#> # A tibble: 333 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804180                 2001              201260 E51B      P803087       
#>  2 P804186                 2001              201260 E51B      P803087       
#>  3 P804189                 2001              201260 E51B      P803087       
#>  4 P804190                 2001              201260 E51B      P803087       
#>  5 P804192                 2001              201260 E51B      P803087       
#>  6 P804548                 2001              201227 E3        K578406       
#>  7 P804549                 2001              201227 E3        K578406       
#>  8 P804550                 2001              201227 E3        K578406       
#>  9 P804552                 2001              201227 E3        K578406       
#> 10 P804553                 2001              201227 E3        K578406       
#> # ℹ 323 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $MP
#> # A tibble: 375 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804101                 2001              201288 MP55      P803016       
#>  2 P804102                 2001              201288 MP55      P803016       
#>  3 P804105                 2001              201288 MP55      P803016       
#>  4 P804109                 2001              201288 MP55      P803016       
#>  5 P804111                 2001              201288 MP55      P803016       
#>  6 P804113                 2001              201288 MP55      P803016       
#>  7 P804114                 2001              201288 MP55      P803016       
#>  8 P804116                 2001              201288 MP55      P803016       
#>  9 P804119                 2001              201288 MP55      P803016       
#> 10 P804121                 2001              201288 MP55      P803016       
#> # ℹ 365 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $O
#> # A tibble: 600 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804142                 2001              201319 O27       P803035       
#>  2 P804143                 2001              201319 O27       P803035       
#>  3 P804144                 2001              201319 O27       P803035       
#>  4 P804145                 2001              201319 O27       P803035       
#>  5 P804146                 2001              201319 O27       P803035       
#>  6 P804181                 2001              201325 O38B      .             
#>  7 P804183                 2001              201325 O38B      .             
#>  8 P804184                 2001              201325 O38B      .             
#>  9 P804185                 2001              201325 O38B      .             
#> 10 P804187                 2001              201325 O38B      .             
#> # ℹ 590 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $P
#> # A tibble: 91 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P805234                 2001              201364 P24       P803249       
#>  2 P805235                 2001              201364 P24       P803249       
#>  3 P805236                 2001              201364 P24       P803249       
#>  4 P805240                 2001              201364 P24       P803249       
#>  5 P805241                 2001              201364 P24       P803249       
#>  6 P805242                 2001              201364 P24       P803249       
#>  7 P805875                 2001              201360 P10       P803358       
#>  8 P805876                 2001              201360 P10       P803358       
#>  9 P805878                 2001              201360 P10       P803358       
#> 10 P805910                 2001              201357 P3        P803362       
#> # ℹ 81 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $SW
#> # A tibble: 405 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804149                 2001              201393 SW113     P803040       
#>  2 P804151                 2001              201393 SW113     P803040       
#>  3 P804152                 2001              201393 SW113     P803040       
#>  4 P804257                 2001              201397 SW128     P803048       
#>  5 P804258                 2001              201397 SW128     P803048       
#>  6 P804259                 2001              201397 SW128     P803048       
#>  7 P804260                 2001              201397 SW128     P803048       
#>  8 P804261                 2001              201397 SW128     P803048       
#>  9 P804262                 2001              201397 SW128     P803048       
#> 10 P804263                 2001              201397 SW128     P803048       
#> # ℹ 395 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#> 
#> $W
#> # A tibble: 384 × 41
#>    chick_ring_number hatch_year hatch_nest_breed_ID hatch_Box hatch_mom_Ring
#>    <chr>                  <dbl>               <dbl> <chr>     <chr>         
#>  1 P804042                 2001              201419 W47       .             
#>  2 P804043                 2001              201417 W43       P803018       
#>  3 P804044                 2001              201417 W43       P803018       
#>  4 P804045                 2001              201417 W43       P803018       
#>  5 P804046                 2001              201419 W47       .             
#>  6 P804047                 2001              201419 W47       .             
#>  7 P804048                 2001              201419 W47       .             
#>  8 P804178                 2001              201414 W37       P803056       
#>  9 P804179                 2001              201414 W37       P803056       
#> 10 P804193                 2001              201428 W60       P803080       
#> # ℹ 374 more rows
#> # ℹ 36 more variables: hatch_nest_dad_Ring <chr>, `Extra-pair_paternity` <chr>,
#> #   `Extra-pair_dad_ring` <chr>, `genetic_dad_ring_(WP_or_EP)` <chr>,
#> #   hatch_nest_LD <chr>, hatch_nest_CS <chr>, hatch_nest_OH <dbl>,
#> #   d0_hatch_nest_brood_size <chr>, d14_hatch_nest_brood_size <chr>,
#> #   rear_nest_breed_ID <dbl>, rear_area <chr>, rear_Box <chr>,
#> #   rear_mom_Ring <chr>, rear_dad_Ring <chr>, rear_nest_trt <dbl>, …
#>