Wrapper function to standardise response variables

This function generates the response data for meta-analysis without standardising the effect sizes / out-of-sample predictions.

Usage

pmap_wrap(..., fns, env = caller_env())

standardise_response(
  data,
  estimate_type = character(1L),
  param_table = NULL,
  dataset = character(1L),
  ...
)

process_response(data, ...)

log_transform_response(data, sim = 10000L, ...)

Arguments

...

Ignored

data

A tibble of analyst data with a list-column called

estimate_type

The type of estimate to be standardised. Character vector of length 1, whose value may be "Zr", "yi", "y25", "y50", "y75".

param_table

A table of estimated 'population' parameters for each variable in the analysis datasets.

dataset

Character vector of length 1. The name of the dataset being processed, e.g. blue tit or eucalyptus.

sim

a numeric vector of length 1L with the number of simulations that should be passed to log_transform()

Value

A tibble of analyst data with standardised values contained in a list-column called 'back_transformed_data'

Functions

standardise_response(): Standardise response data for meta-analysis

process_response(): Process response data for meta-analysis but do not standardise effect-sizes

log_transform_response(): Standardise response data for meta-analysis

standardise_response()

When the estimate_type is "Zr", standardise_response() standardises effect-sizes with est_to_zr(), assuming that the beta_estimate and beta_SE values have already been back-transformed to the appropriate scale. #TODO check this.

When the estimate-type is "yi" or otherwise, the function:

assigns a transformation_type with assign_transformation_type(), assumes that

Converts the out-of-sample predictions on the link- or transformed-response scale back to the original response scale using convert_predictions().

Standardises predictions on the original response-scale to the Z-scale, with pred_to_Z().

Note that for $y_i$ or out of sample predictions that are standardised, if param_table is NA or NULL for a given variable, then the response variable will not be standardised, and NA will be returned for that entry in back_transformed_data.

process_response()

Formats tibbles in the list-column back_transformed_data to ensure that the correct columns are present for meta-analysis, matching the outputs of standardise_response(). For blue tit data data$back_transformed_data$fit and for eucalyptus data, data$back_transformed_data$estimate is renamed Z. se.fit is renamed VZ.

log_transform_response()

maps log_transform_yi() onto back-transformed data stored in list-columns within data

See also

Examples

# Standardise effect-sizes for eucalyptus dataset

data(ManyEcoEvo)
ManyEcoEvo %>%
 filter(dataset == "eucalyptus") %>%
 pluck("data", 1) %>%
 standardise_response(estimate_type = "Zr", 
                      param_table =  NULL,
                      dataset =  "eucalyptus")
#> 
#> ── Computing meta-analysis inputsfor `estimate_type` = "Zr" ────────────────────
#> 
#> ── Computing standardised effect sizes `Zr` and variance `VZr` ──
#> 
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.29212,
#> 3. adjusted_df 21.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.007831,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.07216,
#> 3. adjusted_df 0.560867697.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.57,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.2328,
#> 3. adjusted_df 343.24787.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.3188723,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.0059286,
#> 3. adjusted_df 1.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.007385,
#> 3. adjusted_df 3.5e-25.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.052462,
#> 3. adjusted_df 3.5e-25.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.605,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 8.98,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 7.97,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 5.19,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 18.5,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 5.92,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.529,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 2.89,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.605,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.01312667,
#> 3. adjusted_df -2.6269353.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.197,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.192,
#> 3. adjusted_df 82.703.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.1,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se NA,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.0048042,
#> 3. adjusted_df 341.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.21,
#> 3. adjusted_df NA.
#> ✖ Required values for computing standardised effect sizes missing:
#> ! Returning "NA" for tupple:
#> 1. beta_estimate NA,
#> 2. beta_se 0.5756016,
#> 3. adjusted_df 3.536992.
#> # A tibble: 128 × 40
#>    response_id       submission_id analysis_id split_id TeamIdentifier id_col   
#>    <chr>                     <dbl>       <dbl>    <dbl> <chr>          <chr>    
#>  1 R_0MuZ4chnmwTaLiV             1           1        1 Bicheno        Bicheno-…
#>  2 R_0MuZ4chnmwTaLiV             1           1        2 Bicheno        Bicheno-…
#>  3 R_12R6XuSRUcRemui             1           1        1 Biloela        Biloela-…
#>  4 R_1BWpZlSbkmSofe1             1           1        1 Bingara        Bingara-…
#>  5 R_1Ej3sywrrnzy7KJ             1           1        1 Birchip        Birchip-…
#>  6 R_1FJQlMO4buRu6S2             1           1        1 Birdwoo        Birdwoo-…
#>  7 R_1ib6zw7qoq0lHlf             1           1        1 Blayney        Blayney-…
#>  8 R_1ItMbxJAVt7RRtX             1           1        1 Bodalla        Bodalla-…
#>  9 R_1kH1Ko3yaLsrD0E             1           1        1 Bombala        Bombala-…
#> 10 R_1l9qIgUeSEKY3ME             1           1        1 Bonalbo        Bonalbo-…
#> # ℹ 118 more rows
#> # ℹ 34 more variables: beta_estimate <dbl>, contrast <chr>, adjusted_df <dbl>,
#> #   beta_SE <dbl>, transformation <chr>, link_function_reported <chr>,
#> #   dataset <chr>, mixed_model <dbl>,
#> #   response_transformation_description <chr>,
#> #   response_transformation_status <chr>, response_variable_type <chr>,
#> #   response_construction_description <chr>, response_variable_name <chr>, …
data(ManyEcoEvo_yi)
ManyEcoEvo_yi %>%
filter(dataset == "eucalyptus") %>%
  pluck("data", 1) %>%
  back_transform_response_vars_yi("yi", "eucalyptus") %>%
  log_transform_response()
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for logit-transformed effect sizes or out-of-sample predictions
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for square-root transformed effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for square-root transformed effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for square-root transformed effect sizes or out-of-sample predictions.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> ✔ Applied back-transformation for log-transformed effect sizes or out-of-sample predictions, using 10000 simulations.
#> 
#> ── Computing meta-analysis inputs: ─────────────────────────────────────────────
#> 
#> ── Log-transforming response-variable ──
#> 
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> ✔ Log-transformed out-of-sample predictions, using 10000 simulations.
#> # A tibble: 52 × 26
#>    response_id       submission_id analysis_id split_id response_transformatio…¹
#>    <chr>                     <dbl>       <dbl>    <dbl> <chr>                   
#>  1 R_1ItMbxJAVt7RRtX             1           1        1 NA                      
#>  2 R_1LTZWsikoIaKdMu             1           1        1 NA                      
#>  3 R_1LXqc8NAdABndjl             1           1        1 NA                      
#>  4 R_1NqVW3amqwy69rK             3           3        1 NA                      
#>  5 R_1l9qIgUeSEKY3ME             1           1        1 NA                      
#>  6 R_1l9qIgUeSEKY3ME             2           2        1 NA                      
#>  7 R_1r2FSn2AXuCimpi             1           1        1 NA                      
#>  8 R_1reUKfTYXhTqVMo             1           1        1 NA                      
#>  9 R_21I1NvXiVXIGkqp             1           1        1 NA                      
#> 10 R_21jtSu9hB9Shf6Y             1           1        1 NA                      
#> # ℹ 42 more rows
#> # ℹ abbreviated name: ¹​response_transformation_description
#> # ℹ 21 more variables: response_transformation_status <chr>,
#> #   response_variable_type <chr>, response_construction_description <chr>,
#> #   response_variable_name <chr>, response_id_S2 <chr>, id_col <chr>,
#> #   TeamIdentifier <chr>, question <chr>, file_name <chr>, filepath <chr>,
#> #   checks <named list>, exclusions_all <chr>, mixed_model <dbl>, …