Summarise Analysis Types

Generates a summary of the number of analysis teams, total analyses, models with normal error distributions, mixed effects models, and models developed using Bayesian statistical methods for a given analysis type.

Usage

summarise_analysis_types(ManyEcoEvo_results, ManyEcoEvo_yi_results, ManyEcoEvo)

Arguments

ManyEcoEvo_results

A tibble of ManyEcoEvo_results

ManyEcoEvo_yi_results

A tibble of ManyEcoEvo_yi_results

ManyEcoEvo

A tibble of ManyEcoEvo

Value

A summarised tibble with the variables subset, dataset, num_teams, total_analyses, sum_linear, sum_mixed, sum_Bayesian.

Details

Applies both count_binary_coded_features() and count_teams_analyses() to generate data summaries.

See also

Author

Hannah S. Fraser

Elliot Gould

Examples

summarise_analysis_types(ManyEcoEvo_results, ManyEcoEvo_yi_results, ManyEcoEvo)
#> Adding missing grouping variables: `estimate_type`, `dataset`, `exclusion_set`,
#> `publishable_subset`, `expertise_subset`, `collinearity_subset`
#> Adding missing grouping variables: `estimate_type`, `dataset`, `exclusion_set`
#> Warning: There was 1 warning in `mutate()`.
#> ℹ In argument: `across(...)`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Joining with `by = join_by(subset, dataset)`
#> # A tibble: 4 × 8
#>   subset    dataset total_teams total_analyses sum_linear sum_mixed sum_Bayesian
#>   <chr>     <chr>         <int>          <int>      <dbl>     <dbl>        <dbl>
#> 1 effects   blue t…          63            132        125       129           10
#> 2 effects   eucaly…          40             79         15        62            5
#> 3 predicti… blue t…          43             65         60        64           10
#> 4 predicti… eucaly…           9             19          0        11            2
#> # ℹ 1 more variable: sum_glm <dbl>