Summarise Peer-Reviews

Calculates summary statistics (mean, sd, min, max) of both the number of analyses peer-reviewed by each reviewer, and the number of reviews received by each analysis.

Usage

summarise_reviews(ManyEcoEvo, ManyEcoEvo_results, ManyEcoEvo_yi_results)

Arguments

ManyEcoEvo

A tibble of ManyEcoEvo

ManyEcoEvo_results

A tibble of ManyEcoEvo_results

ManyEcoEvo_yi_results

A tibble of ManyEcoEvo_yi_results

Value

A named list of length two of summarised_data_analyses containing a tibble of summary statistics for each outcome subset (effects or predictions) generated by summarise_analyses_by_reviewer(), and summarised_data_reviews containing a tibble of sumary statistics for each outcome subset, for each dataset generated by summarise_reviews_per_analysis().

See also

Author

Hannah S. Fraser

Elliot Gould

Examples

data(ManyEcoEvo)
data(ManyEcoEvo_results)
data(ManyEcoEvo_yi_results)
summarise_reviews(ManyEcoEvo, ManyEcoEvo_results, ManyEcoEvo_yi_results)
#> Error in group_by(., estimate_type, dataset): Must group by variables found in `.data`.
#> Column `estimate_type` is not found.
#> Column `dataset` is not found.